chore: extract filter label
