+++ ui/apps/ui/src/app/components/filters/filter-date/filter-date.module.ts
@@ -3,10 +3,11 @@ import { CommonModule } from '@angular/common';
 import { FilterDateComponent } from './filter-date.component';
 import { NzDatePickerModule } from 'ng-zorro-antd/date-picker';
 import { FormsModule } from '@angular/forms';
+import {FilterLabelModule} from "@components/filters/filter-label.module";
 
 @NgModule({
   declarations: [FilterDateComponent],
-  imports: [CommonModule, NzDatePickerModule, FormsModule],
+    imports: [CommonModule, NzDatePickerModule, FormsModule, FilterLabelModule],
   exports: [FilterDateComponent],
 })
 export class FilterDateModule {}
