+++ ui/apps/ui/src/app/components/filters/filter-label.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FilterLabelComponent } from '@components/filters/filter-label.component';
+
+@NgModule({
+  declarations: [FilterLabelComponent],
+  imports: [CommonModule],
+  exports: [FilterLabelComponent],
+})
+export class FilterLabelModule {}
