+++ ui/apps/ui/src/app/components/filters/filter-range/filter-range.module.ts
@@ -3,10 +3,11 @@ import { CommonModule } from '@angular/common';
 import { FilterRangeComponent } from './filter-range.component';
 import { NzSliderModule } from 'ng-zorro-antd/slider';
 import { FormsModule } from '@angular/forms';
+import {FilterLabelModule} from "@components/filters/filter-label.module";
 
 @NgModule({
   declarations: [FilterRangeComponent],
-  imports: [CommonModule, NzSliderModule, FormsModule],
+    imports: [CommonModule, NzSliderModule, FormsModule, FilterLabelModule],
   exports: [FilterRangeComponent],
 })
 export class FilterRangeModule {}
